+++ Trabalho-01/pessoal/Veiculo.h
@@ -10,7 +10,7 @@ public:
 
     Veiculo(TIPO t);
     void desenha();
-    void atualiza(float a){}
+    void atualiza(int a){}
     void camera() {}
 };
